update deprecated types
np.bool to np.bool_, np.int to np.int32, np.float to np.float64

+++ ptsa/data/filters/resample.py
@@ -77,7 +77,7 @@ class ResampleFilter(BaseFilter):
 
             # print new_time_axis
 
-            new_time_idx_array = np.rint(new_time_idx_array).astype(np.int)
+            new_time_idx_array = np.rint(new_time_idx_array).astype(np.int32)
 
             new_time_axis = time_axis[new_time_idx_array]
